Initial prototype
This is the state of the prototype that was created at the The Hague sprint, at IPerity.

The code is functional, to the extend that it has been observed to trigger notifications in Conversations. There's not much rhyme or reason to the frequency in which notifications are sent (currently: on any message that is sent to a user of Openfire, that has a non-empty body).