Switch from User Properties to Database
With this change, the code no longer stores Push Notification configuration in user properties. Instead, a new database table is used.

This was primarily motivated by the user property values being limited to 100 characters, while the Push Notification options could be larger (ChatSecure's options appear to be larger).

Note that existing configuration is not migrated.

The version of the plugin has been updated from 0.5.0 to 0.6.0 to denote this change.