OF-3033: Postpone SM termination without push notifications
For a user that has push notifications enabled, a session that is detached (in context of Stream Management) is allowed to be terminated only after a first push notification has been sent.

The rationale for this is that some clients (notably IOS-based clients) are not able to maintain a connection to the server while running in the background (which can happen as often as every 30 seconds). To terminate such clients would lead to a considerable amount of session lifecycles. This implementation considers that such clients would react almost instantly to a push notification. Thus, SM termination is appropriate when a client remains inactive after it was sent the first push notification.

This commit depends on a corresponding change in Openfire.